+++ ft-sdk/src/androidTest/java/com/ft/sdk/tests/RUMTest.java
@@ -12,9 +12,11 @@ import com.ft.sdk.FTResourceEventListener;
 import com.ft.sdk.FTResourceInterceptor;
 import com.ft.sdk.FTSDKConfig;
 import com.ft.sdk.FTSdk;
+import com.ft.sdk.FTExceptionHandler;
 import com.ft.sdk.FTTraceConfig;
 import com.ft.sdk.FTTraceInterceptor;
 import com.ft.sdk.FTTraceManager;
+import com.ft.sdk.FTActivityManager;
 import com.ft.sdk.RUMCacheDiscard;
 import com.ft.sdk.TraceContext;
 import com.ft.sdk.TraceType;
@@ -508,6 +510,61 @@ public class RUMTest extends FTBaseTest {
                 Constants.FT_MEASUREMENT_RUM_ERROR, DataType.RUM_APP, false));
     }
 
+    @Test
+    public void crashErrorForegroundCrashFreeDurationTest() throws Exception {
+        seedCrashFreeTracker(200L, 50L, AppState.RUN, 1_000L);
+
+        FTRUMGlobalManager.get().addError(ERROR, ERROR_MESSAGE, 1_120L, ErrorType.JAVA, AppState.RUN);
+        waitEventConsumeInThreadPool();
+
+        LineProtocolData errorData = getLatestErrorLineProtocolData();
+        Assert.assertEquals(320L, getLongField(errorData, Constants.KEY_RUM_FOREGROUND_CRASH_FREE_DURATION));
+        Assert.assertEquals(50L, getLongField(errorData, Constants.KEY_RUM_BACKGROUND_CRASH_FREE_DURATION));
+
+        Assert.assertEquals(0L, ((Number) Whitebox.getInternalState(FTActivityManager.get(),
+                "foregroundCrashFreeDuration")).longValue());
+        Assert.assertEquals(0L, ((Number) Whitebox.getInternalState(FTActivityManager.get(),
+                "backgroundCrashFreeDuration")).longValue());
+        Assert.assertEquals(1_120L, ((Number) Whitebox.getInternalState(FTActivityManager.get(),
+                "stateStartTime")).longValue());
+    }
+
+    @Test
+    public void crashErrorBackgroundCrashFreeDurationTest() throws Exception {
+        seedCrashFreeTracker(200L, 50L, AppState.BACKGROUND, 1_000L);
+
+        FTRUMGlobalManager.get().addError(ERROR, ERROR_MESSAGE, 1_120L, ErrorType.NATIVE, AppState.BACKGROUND);
+        waitEventConsumeInThreadPool();
+
+        LineProtocolData errorData = getLatestErrorLineProtocolData();
+        Assert.assertEquals(200L, getLongField(errorData, Constants.KEY_RUM_FOREGROUND_CRASH_FREE_DURATION));
+        Assert.assertEquals(170L, getLongField(errorData, Constants.KEY_RUM_BACKGROUND_CRASH_FREE_DURATION));
+    }
+
+    @Test
+    public void preCrashErrorUsesPersistedCrashFreeDurationSnapshotTest() throws Exception {
+        HashMap<String, Object> property = new HashMap<>();
+        property.put(FTExceptionHandler.IS_PRE_CRASH, true);
+
+        clearCrashFreeSharedPreferences();
+        Utils.getSharedPreferences(getContext()).edit()
+                .putLong(Constants.FT_CRASH_FREE_START_TIME, 500L)
+                .putLong(Constants.FT_CRASH_FREE_FOREGROUND_DURATION, 80L)
+                .putLong(Constants.FT_CRASH_FREE_BACKGROUND_DURATION, 20L)
+                .putString(Constants.FT_CRASH_FREE_APP_STATE, AppState.BACKGROUND.toString())
+                .putLong(Constants.FT_CRASH_FREE_STATE_START_TIME, 1_000L)
+                .apply();
+        Whitebox.setInternalState(FTActivityManager.get(), "crashFreeTrackerInitialized", false);
+        Whitebox.setInternalState(FTActivityManager.get(), "pendingCrashSnapshot", (Object) null);
+
+        FTRUMGlobalManager.get().addError(ERROR, ERROR_MESSAGE, 1_120L, ErrorType.NATIVE, AppState.BACKGROUND, property);
+        waitEventConsumeInThreadPool();
+
+        LineProtocolData errorData = getLatestErrorLineProtocolData();
+        Assert.assertEquals(80L, getLongField(errorData, Constants.KEY_RUM_FOREGROUND_CRASH_FREE_DURATION));
+        Assert.assertEquals(140L, getLongField(errorData, Constants.KEY_RUM_BACKGROUND_CRASH_FREE_DURATION));
+    }
+
     /**
      * LongTask data test
      *
@@ -655,6 +712,46 @@ public class RUMTest extends FTBaseTest {
     @Test
     public void traceLinkRUMDataEnable() throws InterruptedException, IOException {
         Assert.assertTrue(checkTraceHasLinkRumData(true));
+    }
+
+    private void seedCrashFreeTracker(long foregroundDuration, long backgroundDuration,
+                                      AppState appState, long stateStartTime) {
+        clearCrashFreeSharedPreferences();
+        Whitebox.setInternalState(FTActivityManager.get(), "crashFreeTrackerInitialized", true);
+        Whitebox.setInternalState(FTActivityManager.get(), "crashFreeStartTime", 0L);
+        Whitebox.setInternalState(FTActivityManager.get(), "foregroundCrashFreeDuration", foregroundDuration);
+        Whitebox.setInternalState(FTActivityManager.get(), "backgroundCrashFreeDuration", backgroundDuration);
+        Whitebox.setInternalState(FTActivityManager.get(), "appState", appState);
+        Whitebox.setInternalState(FTActivityManager.get(), "stateStartTime", stateStartTime);
+        Whitebox.setInternalState(FTActivityManager.get(), "pendingCrashSnapshot", (Object) null);
+    }
+
+    private void clearCrashFreeSharedPreferences() {
+        Utils.getSharedPreferences(getContext()).edit()
+                .remove(Constants.FT_CRASH_FREE_START_TIME)
+                .remove(Constants.FT_CRASH_FREE_FOREGROUND_DURATION)
+                .remove(Constants.FT_CRASH_FREE_BACKGROUND_DURATION)
+                .remove(Constants.FT_CRASH_FREE_APP_STATE)
+                .remove(Constants.FT_CRASH_FREE_STATE_START_TIME)
+                .apply();
+    }
+
+    private LineProtocolData getLatestErrorLineProtocolData() {
+        List<SyncData> recordDataList = FTDBManager.get().queryDataByDataByTypeLimitDesc(0, DataType.RUM_APP);
+        for (SyncData syncData : recordDataList) {
+            LineProtocolData data = new LineProtocolData(syncData.getDataString());
+            if (Constants.FT_MEASUREMENT_RUM_ERROR.equals(data.getMeasurement())) {
+                return data;
+            }
+        }
+        Assert.fail("No rum error data found");
+        return null;
+    }
+
+    private long getLongField(LineProtocolData data, String key) {
+        Object value = data.getField(key);
+        Assert.assertNotNull(key + " should exist", value);
+        return ((Number) value).longValue();
     }
 
     /**